build: land bullet + ui_gameinfo_mp (~112 files)
game/bullet.cpp: 2 paren fixes (friendlyfire OnSameTeam,
revTraceHit allsolid startsolid) + 1 [[maybe_unused]] on hitEntId.

ui_mp/ui_gameinfo_mp.cpp: 2 KISAKHACK uintptr_t bridges (Material*
slot in serverHardwareIconList[] holding bit-packed flags, FS_GetFileList
return-as-pointer in v1/v8 comparison loop).

posix_backbone_stubs.cpp additions (~17 entries):
- BG_AdvanceTrace (char), BG_GetSurfacePenetrationDepth (double),
  BG_GetWeaponIndex (unsigned int).
- DirToByte (unsigned char), FS_GetFileList (int).
- G_AntiLag_RestoreClientPos, G_AntiLagRewindClientPos,
  G_CheckHitTriggerDamage, G_Damage, G_LocationalTraceAllowChildren,
  G_TempEntity (gentity_s*), OnSameTeam (bool).
- bullet_penetrationEnabled, bullet_penetrationMinFxDist,
  g_debugLocDamage, sv_clientSideBullets dvar* globals.
- sharedUiInfo (sharedUiInfo_t, typed from ui_mp.h).

+++ src/ui_mp/ui_gameinfo_mp.cpp
@@ -114,8 +114,11 @@ void __cdecl UI_LoadArenas()
                 for (i = 0; i < sharedUiInfo.numGameTypes; ++i)
                 {
                     if (!I_stricmp(pszToken, sharedUiInfo.gameTypes[i].gameType))
-                        sharedUiInfo.serverHardwareIconList[40 * sharedUiInfo.mapCount - 5115] = (Material *)((int)sharedUiInfo.serverHardwareIconList[40 * sharedUiInfo.mapCount - 5115]
-                            | (1 << i));
+                        // KISAKHACK: serverHardwareIconList holds Material* slots
+                        // bit-packed with flags. Bridge through uintptr_t.
+                        sharedUiInfo.serverHardwareIconList[40 * sharedUiInfo.mapCount - 5115] =
+                            (Material *)((uintptr_t)sharedUiInfo.serverHardwareIconList[40 * sharedUiInfo.mapCount - 5115]
+                                | (1 << i));
                 }
             }
             Com_EndParseSession();
@@ -144,11 +147,13 @@ const char *UI_LoadArenasFromFile_LoadObj()
     unsigned int v9; // [esp+24B4h] [ebp-4h]
 
     ui_numArenas = 0;
-    result = (char *)FS_GetFileList("mp", "arena", FS_LIST_PURE_ONLY, listbuf, 1024);
+    // FS_GetFileList returns int (count); hex-rays decompiled as char*.
+    // Bridge through uintptr_t.
+    result = (char *)(uintptr_t)FS_GetFileList("mp", "arena", FS_LIST_PURE_ONLY, listbuf, 1024);
     v1 = result;
     v3 = listbuf;
     v8 = 0;
-    while (v8 < (int)v1)
+    while (v8 < (int)(uintptr_t)v1)
     {
         v9 = strlen(v3);
         snprintf(string, ARRAYSIZE(string), "%s/%s", "mp", v3);
